feat: get ground truth variant list
